# ui/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ArtifactBody": {
        "properties": {
          "data": {
            "type": "object"
          },
          "key": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "type": "string"
          },
          "run_id": {
            "format": "int64",
            "type": "integer"
          },
          "task_run_id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "run_id",
          "kind",
          "data"
        ],
        "type": "object"
      },
...
      "ArtifactsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtifactListItem"
            },
            "type": "array"
          },
          "next_cursor": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "PrefilterBody": {
        "properties": {
          "skip": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "skip"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first pipeline orchestrator API",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "cereyan",
    "version": "1.14.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/artifacts": {
      "get": {
        "operationId": "list_artifacts",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "key",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "flow",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Keyset cursor: the id of the last artifact of the previous page.",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactsPage"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "observability"
        ]
      },
      "post": {
        "operationId": "create_artifact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRow"
                }
              }
            },
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "tags": [
          "observability"
        ]
      }
    },
    "/api/backfills/{id}": {
      "get": {
        "operationId": "get_backfill",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackfillStatus"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "backfills"
        ]
      }
    },
    "/api/backfills/{id}/cancel": {
      "post": {
        "operationId": "cancel_backfill",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackfillStatus"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "backfills"
        ]
      }
    },
    "/api/backfills/{id}/prefilter": {
      "post": {
        "operationId": "prefilter_backfill",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrefilterBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "backfills"
        ]
      }
    },
    "/api/counts": {
      "get": {
        "operationId": "counts",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Counts"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "counts"
        ]
      }
    },
...
    "/api/database/reset": {
      "post": {
        "operationId": "reset_database",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetResult"
                }
              }
            },
            "description": ""
          },
          "409": {
            "description": "A reset is already running"
          },
          "422": {
            "description": "Missing or unknown scope"
          },
          "500": {
            "description": "The copy failed; nothing was deleted"
          }
        },
        "tags": [
          "database"
        ]
      }
    },
    "/api/engine/failed": {
      "post": {
        "operationId": "failed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Queued runs of the module failed"
          }
        },
        "tags": [
          "engine"
        ]
      }
    },
    "/api/engine/heartbeat": {
      "post": {
        "operationId": "heartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "{cancel: bool, active: bool}"
          }
        },
        "tags": [
          "engine"
        ]
      }
    },
    "/api/engine/report": {
      "post": {
        "operationId": "report",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "engine"
        ]
      }
    },
    "/api/engine/work": {
      "post": {
        "operationId": "work",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "engine"
        ]
      }
    },
    "/api/events": {
      "get": {
        "operationId": "list_events",
        "parameters": [
          {
            "description": "Exact name, or a prefix ending in `*` (e.g. `run.*`).",
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "resource_kind",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "resource_id",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "flow_id",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Keyset cursor: the id of the last event of the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Ascending order when true (defaults to newest first).",
            "in": "query",
            "name": "ascending",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsPage"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "observability"
        ]
      },
      "post": {
        "operationId": "emit_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmitEventBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "observability"
        ]
      }
    },
    "/api/events/{id}": {
      "get": {
        "operationId": "get_event",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "observability"
        ]
      }
    },
    "/api/flows": {
      "get": {
        "operationId": "list_flows",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FlowSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "flows"
        ]
      }
    },
    "/api/flows/{id}": {
      "delete": {
        "operationId": "delete_flow",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Flow is live"
          }
        },
        "tags": [
          "flows"
        ]
      },
      "get": {
        "operationId": "get_flow",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowSummary"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "flows"
        ]
      }
    },
    "/api/flows/{id}/backfill": {
      "post": {
        "operationId": "create_backfill",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackfillStatus"
                }
              }
            },
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "tags": [
          "backfills"
        ]
      }
    },
    "/api/flows/{id}/backfills": {
      "get": {
        "operationId": "list_backfills",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BackfillStatus"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "backfills"
        ]
      }
    },
    "/api/flows/{id}/pause": {
      "post": {
        "operationId": "pause_flow",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All schedules paused"
          }
        },
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/flows/{id}/resume": {
      "post": {
        "operationId": "resume_flow",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All schedules resumed"
          }
        },
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/flows/{id}/runs": {
      "post": {
        "operationId": "create_run_for_flow",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunForFlowBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": ""
          },
          "422": {
            "description": "Invalid parameters"
          }
        },
        "tags": [
          "flows"
        ]
      }
    },
    "/api/flows/{id}/schedules": {
      "get": {
        "operationId": "list_schedules",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleRow"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "operationId": "create_schedule",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRow"
                }
              }
            },
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/flows/{id}/upcoming": {
      "get": {
        "operationId": "upcoming_runs",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Also list this many fires of each active schedule past its\nmaterialized runs, computed without creating runs (at most 100).",
            "in": "query",
            "name": "projected",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UpcomingItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "schedules"
        ]
      }
    },
...
    "/api/resources/acquire": {
      "post": {
        "operationId": "acquire",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcquireRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "{lease: id} or {waiting: name}"
          }
        },
        "tags": [
          "engine"
        ]
      }
    },
    "/api/resources/release": {
      "post": {
        "operationId": "release",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "engine"
        ]
      }
    },
...
  }
}